LayoutTests/platform/ios-sim-deprecated/fast/scrolling/scrollbar-mousedown-move-mouseup-expected.txt

This is a scrollable div.

PASS events['scrollme'].length is 4
PASS events['notscrollme'].length is 0
PASS events['scrollme'][0].type is "mousedown"
PASS events['scrollme'][0].which is 1
PASS events['scrollme'][1].type is "mouseup"
PASS events['scrollme'][1].which is 1
PASS events['scrollme'][2].type is "mousedown"
FAIL events['scrollme'][2].which should be 2. Was 1.
PASS events['scrollme'][3].type is "mouseup"
FAIL events['scrollme'][3].which should be 2. Was 1.
PASS successfullyParsed is true

TEST COMPLETE
